+++ src/theme/components/TooltipTheme.ts
@@ -1,0 +1,11 @@
+import { Components } from "@mui/material"
+
+export default function TooltipTheme(): Components {
+  return {
+    MuiTooltip: {
+      defaultProps: {
+        arrow: true,
+      },
+    },
+  }
+}
